Fix initialization of empty collections
