+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/collection/internal/DelayedCollectionInitializer.java
@@ -12,102 +12,157 @@ import org.hibernate.engine.spi.CollectionKey;
 import org.hibernate.engine.spi.PersistenceContext;
 import org.hibernate.engine.spi.SharedSessionContractImplementor;
 import org.hibernate.internal.log.LoggingHelper;
-import org.hibernate.metamodel.CollectionClassification;
 import org.hibernate.metamodel.mapping.PluralAttributeMapping;
 import org.hibernate.persister.collection.CollectionPersister;
 import org.hibernate.query.NavigablePath;
-import org.hibernate.sql.results.graph.DomainResultAssembler;
 import org.hibernate.sql.results.graph.FetchParentAccess;
+import org.hibernate.sql.results.graph.Initializer;
+import org.hibernate.sql.results.graph.collection.CollectionInitializer;
+import org.hibernate.sql.results.graph.collection.LoadingCollectionEntry;
+import org.hibernate.sql.results.graph.entity.EntityInitializer;
 import org.hibernate.sql.results.jdbc.spi.RowProcessingState;
 
 /**
  * @author Steve Ebersole
  */
-public class DelayedCollectionInitializer extends AbstractCollectionInitializer {
+public class DelayedCollectionInitializer implements CollectionInitializer {
 
-	// per-row state
-	private PersistentCollection collectionInstance;
+	private final NavigablePath fetchedPath;
+	private final PluralAttributeMapping fetchedMapping;
+	private final FetchParentAccess parentAccess;
+
+	private CollectionKey collectionKey;
+	private PersistentCollection instance;
 
 	public DelayedCollectionInitializer(
-			NavigablePath navigablePath,
-			PluralAttributeMapping attributeMapping,
-			FetchParentAccess parentAccess,
-			DomainResultAssembler keyContainerAssembler,
-			DomainResultAssembler keyCollectionAssembler) {
-		super( navigablePath, attributeMapping, parentAccess, false, keyContainerAssembler, keyCollectionAssembler );
+			NavigablePath fetchedPath,
+			PluralAttributeMapping fetchedMapping,
+			FetchParentAccess parentAccess) {
+		this.fetchedPath = fetchedPath;
+		this.fetchedMapping = fetchedMapping;
+		this.parentAccess = parentAccess;
 	}
 
 	@Override
-	public PersistentCollection getCollectionInstance() {
-		return collectionInstance;
+	public NavigablePath getNavigablePath() {
+		return fetchedPath;
 	}
 
 	@Override
 	public void resolveKey(RowProcessingState rowProcessingState) {
-		super.resolveKey( rowProcessingState );
+		if ( collectionKey != null ) {
+			// already resolved
+			return;
+		}
 
-		getParentAccess().registerResolutionListener(
-				owner -> collectionInstance.setOwner( owner )
-		);
-	}
+		final Object parentKey = parentAccess.getParentKey();
+		if ( parentKey != null ) {
+			collectionKey = new CollectionKey(
+					fetchedMapping.getCollectionDescriptor(),
+					parentKey
+			);
 
-	@Override
-	public void resolveInstance(RowProcessingState rowProcessingState) {
-		final CollectionKey collectionKey = resolveCollectionKey( rowProcessingState );
-		if(collectionKey != null) {
-
-			final SharedSessionContractImplementor session = rowProcessingState.getSession();
-			final PersistenceContext persistenceContext = session.getPersistenceContext();
-
-			final PluralAttributeMapping attributeMapping = getCollectionAttributeMapping();
-			final CollectionPersister collectionDescriptor = attributeMapping.getCollectionDescriptor();
-
-			// todo (6.0) : look for LoadingCollectionEntry?
-
-			final PersistentCollection existing = persistenceContext.getCollection( collectionKey );
-			if ( existing != null ) {
-				collectionInstance = existing;
-			}
-			else {
-				final CollectionSemantics collectionSemantics = collectionDescriptor.getCollectionSemantics();
-
-				collectionInstance = collectionSemantics.instantiateWrapper(
-						collectionKey.getKey(),
-						collectionDescriptor,
-						session
-				);
-
-				getParentAccess().registerResolutionListener(
-						owner -> collectionInstance.setOwner( owner )
-				);
-
-				persistenceContext.addUninitializedCollection(
-						collectionDescriptor,
-						collectionInstance,
-						collectionKey.getKey()
-				);
-
-				if ( collectionSemantics.getCollectionClassification() == CollectionClassification.ARRAY ) {
-					session.getPersistenceContext().addCollectionHolder( collectionInstance );
-				}
-			}
+			parentAccess.registerResolutionListener( owner -> instance.setOwner( owner ) );
 		}
 	}
 
 	@Override
-	public void initializeInstance(RowProcessingState rowProcessingState) {
-		/// nothing to do
+	public void resolveInstance(RowProcessingState rowProcessingState) {
+		final SharedSessionContractImplementor session = rowProcessingState.getSession();
+		final PersistenceContext persistenceContext = session.getPersistenceContext();
+
+		if ( collectionKey != null ) {
+			EntityInitializer entityInitializer = getEntityInitializer( rowProcessingState );
+
+
+			final Object entityUsingInterceptor = persistenceContext.getEntity( entityInitializer.getEntityKey() );
+			if ( entityUsingInterceptor != null ) {
+				return;
+			}
+
+			final Object key = collectionKey.getKey();
+
+			final LoadingCollectionEntry loadingEntry = persistenceContext.getLoadContexts()
+					.findLoadingCollectionEntry( collectionKey );
+			final PersistentCollection registeredInstance = persistenceContext.getCollection( collectionKey );
+
+			if ( loadingEntry != null ) {
+				instance = loadingEntry.getCollectionInstance();
+				return;
+			}
+
+			if ( registeredInstance != null ) {
+				instance = registeredInstance;
+				return;
+			}
+
+			instance = makePersistentCollection( fetchedMapping, key, rowProcessingState );
+
+			persistenceContext.addUninitializedCollection(
+					getInitializingCollectionDescriptor(),
+					instance,
+					key
+			);
+		}
+		else {
+			instance = makePersistentCollection( fetchedMapping, collectionKey, rowProcessingState );
+			instance.initializeEmptyCollection( getInitializingCollectionDescriptor() );
+			persistenceContext.addNonLazyCollection( instance );
+		}
 	}
 
-	@Override
-	public void finishUpRow(RowProcessingState rowProcessingState) {
-		collectionInstance = null;
+	private static PersistentCollection makePersistentCollection(
+			PluralAttributeMapping fetchedMapping,
+			Object collectionKey,
+			RowProcessingState rowProcessingState) {
+		final CollectionPersister collectionDescriptor = fetchedMapping.getCollectionDescriptor();
+		final CollectionSemantics collectionSemantics = collectionDescriptor.getCollectionSemantics();
 
-		super.finishUpRow( rowProcessingState );
+		return collectionSemantics.instantiateWrapper(
+				collectionKey,
+				collectionDescriptor,
+				rowProcessingState.getSession()
+		);
+	}
+
+	private EntityInitializer getEntityInitializer(RowProcessingState rowProcessingState) {
+		Initializer initializer = rowProcessingState.resolveInitializer( getNavigablePath().getParent() );
+		while ( !( initializer instanceof EntityInitializer ) ) {
+			initializer = rowProcessingState.resolveInitializer( initializer.getNavigablePath().getParent() );
+		}
+		return (EntityInitializer) initializer;
+
+	}
+
+
+	@Override
+	public void initializeInstance(RowProcessingState rowProcessingState) {
 	}
 
 	@Override
 	public String toString() {
 		return "DelayedCollectionInitializer(" + LoggingHelper.toLoggableString( getNavigablePath() ) + ")";
+	}
+
+	@Override
+	public void finishUpRow(RowProcessingState rowProcessingState) {
+		collectionKey = null;
+		instance = null;
+	}
+
+	@Override
+	public PluralAttributeMapping getInitializedPart() {
+		return fetchedMapping;
+	}
+
+	@Override
+	public PersistentCollection getCollectionInstance() {
+		return instance;
+	}
+
+	@Override
+	public CollectionKey resolveCollectionKey(RowProcessingState rowProcessingState) {
+		resolveKey( rowProcessingState );
+		return collectionKey;
 	}
 }
